+++ lib/attributor/types/date_time.rb
@@ -19,7 +19,7 @@ module Attributor
       def self.load(value,context=Attributor::DEFAULT_ROOT_CONTEXT, **options)
         # We assume that if the value is already in the right type, we've decoded it already
         return value if value.is_a?(self.native_type)
-        return value.to_datetime if value.is_a?(::Time)
+        return value.to_datetime if value.respond_to?(:to_datetime)
         return nil unless value.is_a?(::String)
         # TODO: we should be able to convert not only from String but Time...etc
         # Else, we'll decode it from String.